pyproject.toml
[tool.poetry]
name = "movie-recommendation-system"
version = "0.1.0"
description = "Recommendation System project for LLM course"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
transformers = "^4.48.1"
sqlalchemy = "^2.0.37"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
